pyproject.toml
[project]
name = "dailystockstory"
version = "0.1.0"
description = "Stock briefing data collector and reader"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "yfinance>=0.2.36",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pillow>=12.1.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
]
